requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
opencv-python==4.8.1.78
librosa==0.10.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
mediapipe==0.10.3
pytest==7.4.2
pytest-cov==4.1.0
gradio==3.50.2
Pillow==10.0.1
soundfile==0.12.1
joblib==1.3.2
shap==0.42.1
lime==0.2.0.1